requirements.txt
absl-py==0.6.1
astor==0.7.1
astroid==1.6.5
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
bleach==1.5.0
cairocffi==0.9.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.5.0
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.7
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
gizeh==0.1.11
grpcio==1.17.1
h5py==2.8.0
html5lib==0.9999999
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonlines==1.2.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
more-itertools==4.3.0
numpy==1.15.4
pathlib==1.0.1
pathlib2==2.3.3
pbr==5.1.1
Pillow==5.3.0
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pylint==1.9.3
pypng==0.0.18
pytest==4.0.2
python-dotenv==0.10.1
requests==2.21.0
scandir==1.9.0
singledispatch==3.4.0.3
six==1.12.0
tensorboard==1.12.1
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
wrapt==1.10.11
